Create ApiController.groovy
This is the only function which is being called as both Portfolio and Folder items are saved. 

In summary, as a Folder item is saved, this script calls Portfolio's expert script which updates Portfolio's attachment, i.e. Folder items, list. On the other hand, as a Portfolio item is save, it calls Folder's expert script which copy down the reviewers information from its Portfolio item.